Revert "Natively support Launch Darkly (#117)"
This reverts commit 9072f5aa616465333f8d8ccf9260c5a49a169be3.

+++ integrations/launch-darkly/src/commonTest/kotlin/io/github/kevincianfarini/monarch/launchdarkly/MutableLDClientInterface.kt
@@ -4,6 +4,6 @@ interface MutableLDClientInterface {
     fun setVariation(flagKey: String, value: Boolean)
     fun setVariation(flagKey: String, value: String)
     fun setVariation(flagKey: String, value: Double)
-    fun setVariation(flagKey: String, value: Long)
+    fun setVariation(flagKey: String, value: Int)
 }
